Rename librad2 to librad
We have an existing `rad` executable, which we don't want to clash with,
but `librad` is still free. Opens up future version evolution via
`librad2` to `librad++11`.
