+++ librad/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "librad2"
+name = "librad"
 version = "0.1.0"
 authors = ["Monadic <team@monadic.xyz>"]
 edition = "2018"
