+++ librad/src/meta/common.rs
